ab1: code optimization and reorganization
