+++ front-end/src/features/login.js
@@ -66,6 +66,8 @@ const { actions, reducer } = createSlice({
   },
 });
 
+export default reducer;
+
 export function fetchToken(email, password) {
   return async (dispatch, getState) => {
     const status = loginSelector(getState()).status;
@@ -114,5 +116,3 @@ export function logoutProfile() {
     dispatch(actions.logouting());
   };
 }
-
-export default reducer;
