+++ front-end/src/utils/store.js
@@ -1,5 +1,5 @@
 import loginReducer from "../features/login";
-import userReducer from "../features/signup";
+import userReducer from "../features/user";
 import { configureStore } from "@reduxjs/toolkit";
 
 export default configureStore({
